requirements.txt
django
djangorestframework
django-filter
mysqlclient
python-dotenv
